# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Message API",
        "version": "0.1"
    },
    "paths": {
        "/messages/send": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message",
                "operationId": "cbf97e68f247c0084e0c7a9f112e135b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_number",
                                    "content"
                                ],
                                "properties": {
                                    "receiver_number": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/messages/receive/{receiver_number}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Receive a message",
                "operationId": "77d7cd5cb69ffdeb836c7d5c39c8c5dd",
                "parameters": [
                    {
                        "name": "receiver_number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": "['message1', 'message2']"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}